moved adminrouter to the routes folder
